fix test_auto_gptq.py

+++ auto_round/export/export_to_autogptq/export.py
@@ -184,7 +184,7 @@ def pack_layer(name, model, backend, device=None):
     if param_count == 2:
         qlayer.pack(layer, scale, device)
     else:
-        qlayer.pack(layer, scale, zero, None, device)
+        qlayer.pack(layer, scale, zero.float(), None, device)
     qlayer.to(orig_device)
     set_module(model, name, qlayer)
     # Note: release weight and bias explicitly, in case they are referenced elsewhere
